chore(clickable-tile): refactored and updated props
